+++ dbt_coves/config/config.py
@@ -25,8 +25,32 @@ class GenerateModel(BaseModel):
     sources: Optional[GenerateSourcesModel] = GenerateSourcesModel()
 
 
+class ExtractAirbyteModel(BaseModel):
+    path: Optional[str] = ""
+    host: Optional[str] = ""
+    port: Optional[str] = ""
+    dbt_list_args: Optional[str] = ""
+
+
+class LoadAirbyteModel(BaseModel):
+    path: Optional[str] = ""
+    host: Optional[str] = ""
+    port: Optional[str] = ""
+    secrets: Optional[str] = ""
+
+
+class ExtractModel(BaseModel):
+    airbyte: Optional[ExtractAirbyteModel] = ExtractAirbyteModel()
+
+
+class LoadModel(BaseModel):
+    airbyte: Optional[LoadAirbyteModel] = LoadAirbyteModel()
+
+
 class ConfigModel(BaseModel):
     generate: Optional[GenerateModel] = GenerateModel()
+    extract: Optional[ExtractModel] = ExtractModel()
+    load: Optional[LoadModel] = LoadModel()
 
 
 class DbtCovesConfig:
@@ -41,6 +65,14 @@ class DbtCovesConfig:
         "generate.sources.model_props_strategy",
         "generate.sources.templates_folder",
         "generate.sources.metadata",
+        "extract.airbyte.path",
+        "extract.airbyte.host",
+        "extract.airbyte.port",
+        "extract.airbyte.dbt_list_args",
+        "load.airbyte.path",
+        "load.airbyte.host",
+        "load.airbyte.port",
+        "load.airbyte.secrets",
     ]
 
     def __init__(self, flags: DbtCovesFlags) -> None: