+++ dbt_coves/tasks/check.py
@@ -13,7 +13,7 @@ class CheckTask(BaseConfiguredTask):
     """
     Task that runs pre-commit and sqlfluff
     """
-    
+
     @classmethod
     def register_parser(cls, sub_parsers, base_subparser):
         subparser = sub_parsers.add_parser(
@@ -29,18 +29,23 @@ class CheckTask(BaseConfiguredTask):
         return subparser
 
     def run(self) -> int:
-        console.print("Running pre-commit hooks on staged and commmitted git files...\n")
+        console.print(
+            "Running pre-commit hooks on staged and commmitted git files...\n"
+        )
 
         command = shell_run(["pre-commit", "run", "-a"])
         if command.returncode != 0:
             return command.returncode
 
+        sql_fluff_status = 0
         for source_path in self.config.source_paths:
             console.print(f"Linting files in [u]{source_path}[/u]...\n")
 
             command = shell_run(["sqlfluff", "lint", source_path])
 
-            if not self.coves_flags.check["no-fix"] and command.returncode != 0:
+            if command.returncode != 0:
+                sql_fluff_status = command.returncode
+            if not self.coves_flags.check["no-fix"] and sql_fluff_status != 0:
                 confirmed = questionary.confirm(
                     f"Would you like to try auto-fixing linting errors in {source_path}?",
                     default=True,
@@ -49,4 +54,4 @@ class CheckTask(BaseConfiguredTask):
                     command = fix(source_path)
                     if command.returncode != 0:
                         return command.returncode
-        return 0
+        return sql_fluff_status