+++ dbt_coves/utils/shell.py
@@ -1,4 +1,5 @@
 from subprocess import PIPE, Popen, run as shell_run
+import subprocess
 
 
 def run(cmd, cwd=None):
@@ -10,3 +11,16 @@ def run(cmd, cwd=None):
 
 def run_and_capture(args_list):
     return shell_run(args_list, capture_output=True, text=True)
+
+
+def run_dbt_ls(bash_cmd, cwd=None):
+    process = subprocess.run(
+        bash_cmd.split(),
+        cwd=cwd,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,
+        check=True,
+    )
+    stdout = process.stdout.decode().strip()
+    full_stdout = stdout.split("\n") if "\n" in stdout else [stdout]
+    return [line for line in full_stdout if "source:" in line]